dicom-mwl-mpps/docker/mwl-api/requirements.txt
fastapi
uvicorn
pydicom
mysql-connector-python
python-dotenv
jinja2
python-multipart